Add next artwork and open details notification actions
For sources that support the 'Next Artwork' built in action (BUILTIN_COMMAND_ID_NEXT_ARTWORK), we now show a 'Next' action in the notification so that users can skip to the next artwork in response to a notification.

For sources that include a viewIntent in their artwork, we now show a 'Details' action which directly links to this intent.